Integracion de notificaciones en todos los modulos

+++ CERVERICA/DTO/Recetas/PasosRecetaDto.cs
@@ -4,6 +4,6 @@
     {
         public string Descripcion { get; set; }
         public int Orden { get; set; }
-        public int Tiempo { get; set; }
+        public double Tiempo { get; set; }
     }
 }
